+++ paddle/phi/kernels/cpu/activation_grad_kernel.cc
@@ -13,10 +13,11 @@ See the License for the specific language governing permissions and
 limitations under the License. */
 
 #include "paddle/phi/kernels/activation_grad_kernel.h"
+
 #include "paddle/phi/backends/cpu/cpu_context.h"
 #include "paddle/phi/common/float16.h"
 #include "paddle/phi/core/kernel_registry.h"
-#include "paddle/phi/kernels/impl/activation_impl.h"
+#include "paddle/phi/kernels/impl/activation_grad_impl.h"
 
 namespace phi {
